Squashed patch containing all changes for 7.1:
Fix for all Streetcar arrivals - the data provider changed the data without any warning (the data doesn’t come from TriMet).

Also:
- Added back ‘Bookmarks at the top’ setting.
- iOS 7: Fixed refresh timer button so it doesn’t flash.
- iOS 7: Fixed some graphics glitches.
- Canceled busses show the time crossed out.
- Fixed app store link from main screen.
- Now uses Apple’s geocoder for getting addresses from GPS locations.
- Fixed issues when GPS is used for destination in Trip Planner.
- Added back proximity alarm icon.
- Removed black color theme; it does not work well in iOS7.
- Changed JELD-WEN Field to Providence Park. Go Timbers!
- Added route map with stops to arrival details.
- Toolbar map icon now behaves consistantly added new menu item to map arrivals
- Added new URL scheme to allow other apps to launch PDX Bus’s trip planner.
- Icon has been tweaked.
- Improved responsiveness, memory usage and display of MAX & Streetcar maps.
- Minor fixes to station search.
- The ‘Nearby stops’ option for arrivals has been updated to to show maps, routes or arrivals.

+++ src/ZXing/cpp/core/src/zxing/aztec/decoder/Decoder.cpp
@@ -403,7 +403,7 @@ Ref<BitArray> Decoder::extractBits(Ref<zxing::BitMatrix> matrix) {
   }
             
   int layer = ddata_->getNBLayers();
-  int size = matrix->getHeight();
+  int size = (int)matrix->getHeight();
   int rawbitsOffset = 0;
   int matrixOffset = 0;
             
@@ -443,7 +443,7 @@ Ref<BitArray> Decoder::extractBits(Ref<zxing::BitMatrix> matrix) {
 }
         
 Ref<BitMatrix> Decoder::removeDashedLines(Ref<zxing::BitMatrix> matrix) {
-  int nbDashed = 1 + 2 * ((matrix->getWidth() - 1) / 2 / 16);
+  int nbDashed = 1 + 2 * (((int)matrix->getWidth() - 1) / 2 / 16);
   Ref<BitMatrix> newMatrix(new BitMatrix(matrix->getWidth() - nbDashed, matrix->getHeight() - nbDashed));
             
   int nx = 0;
