+++ src/ZXing/cpp/core/src/zxing/common/BitSource.cpp
@@ -72,6 +72,6 @@ int BitSource::readBits(int numBits) {
 }
 
 int BitSource::available() {
-  return 8 * (bytes_.size() - byteOffset_) - bitOffset_;
+  return (int)(8 * (bytes_.size() - byteOffset_) - bitOffset_);
 }
 }
