+++ src/ZXing/cpp/core/src/zxing/datamatrix/detector/Detector.cpp
@@ -409,7 +409,7 @@ Ref<BitMatrix> Detector::sampleGrid(Ref<BitMatrix> image, int dimensionX, int di
 
 void Detector::insertionSort(std::vector<Ref<ResultPointsAndTransitions> > &vector) {
 #ifndef __clang_analyzer__
-  int max = vector.size();
+  int max = (int)vector.size();
   bool swapped = true;
   Ref<ResultPointsAndTransitions> value;
   Ref<ResultPointsAndTransitions> valueB;
