+++ src/ZXing/cpp/core/src/zxing/oned/EAN13Reader.cpp
@@ -33,7 +33,7 @@ namespace zxing {
       const int countersLen = 4;
       int counters[countersLen] = { 0, 0, 0, 0 };
 
-      int end = row->getSize();
+      int end = (int)row->getSize();
       int rowOffset = startGuardEnd;
       int lgPatternFound = 0;
