+++ src/ZXing/cpp/core/src/zxing/oned/EAN8Reader.cpp
@@ -29,7 +29,7 @@ namespace zxing {
       const int countersLen = 4;
       int counters[countersLen] = { 0, 0, 0, 0 };
 
-      int end = row->getSize();
+      int end = (int)row->getSize();
       int rowOffset = startGuardEnd;
 
       for (int x = 0; x < 4 && rowOffset < end; x++) {
