+++ src/ZXing/cpp/core/src/zxing/qrcode/decoder/DecodedBitStreamParser.cpp
@@ -73,7 +73,7 @@ void DecodedBitStreamParser::append(std::string &result,
     return;
   }
 
-  const int maxOut = 4 * nIn + 1;
+  const int maxOut = (int)(4 * nIn + 1);
   unsigned char* bufOut = new unsigned char[maxOut];
 
   ICONV_CONST char *fromPtr = (ICONV_CONST char *)bufIn;
@@ -91,7 +91,7 @@ void DecodedBitStreamParser::append(std::string &result,
   }
   iconv_close(cd);
 
-  int nResult = maxOut - nTo;
+  int nResult = maxOut - (int)nTo;
   bufOut[nResult] = '\0';
   result.append((const char *)bufOut);
   delete[] bufOut;
